Fix startup import cycle and apply options settings
