+++ custom_components/media_cover_art/models.py
@@ -1,0 +1,19 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+
+@dataclass(slots=True)
+class TrackQuery:
+    artist: str | None
+    title: str | None
+    album: str | None
+    artwork_size: int
+
+
+@dataclass(slots=True)
+class ResolvedCover:
+    provider: str
+    artwork_url: str | None
+    content_type: str
+    image: bytes
